DSM telemetry flightlog (issue #590 field mapping bugfix and max "9999" number enhancement) - fixed bug #590: overwriting struct elements in flightlog protocol mapping from read packet - change numbers from max "255" to "9999" for 7-8CH+ RX, 200SRX/180CFX gain tweaking   (see issue #571: https://bitbucket.org/PhracturedBlue/deviation/issue/571/telemetry-only-partially-functional-on-200,    see DeviationTX forum: http://www.deviationtx.com/forum/protocol-development/1192-dsm-telemetry-support?start=100#26681)

+++ src/telemetry.h
@@ -128,9 +128,9 @@ struct telem_devo {
 };
 struct telem_dsm_flog {
     //Do not change the order of these, they are aligned to the dsm packet
-    u8 fades[4];
-    u8 frameloss;
-    u8 holds;
+    u16 fades[4];
+    u16 frameloss;
+    u16 holds;
     u16 volt[2];
     u16 rpm;
     s16 temp;
